+++ src/Metrics/Metrics.php
@@ -10,7 +10,7 @@ use Sentry\Metrics\Types\DistributionType;
 use Sentry\Metrics\Types\GaugeType;
 use Sentry\Metrics\Types\SetType;
 
-final class Metrics
+class Metrics
 {
     /**
      * @var self|null
@@ -131,20 +131,20 @@ final class Metrics
     /**
      * @template T
      *
-     * @param callable(): T         $callable
+     * @param callable(): T         $callback
      * @param array<string, string> $tags
      *
      * @return T
      */
     public function timing(
         string $key,
-        callable $callable,
+        callable $callback,
         array $tags = [],
         int $stackLevel = 0
     ) {
         $startTimestamp = microtime(true);
 
-        $result = $callable();
+        $result = $callback();
 
         $this->aggregator->add(
             DistributionType::TYPE,